requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
jdatetime==4.1.0
persiantools==3.0.1
pika==1.2.1
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
typing_extensions==4.2.0
urllib3==1.26.9
